phoGPT/train/requirements.txt
# Data handling and datasets
datasets>=2.14.0
transformers>=4.46.0
einops>=0.6.0

# Machine Learning and Deep Learning
torch>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.5.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Training and fine-tuning
accelerate>=0.35.0
peft>=0.13.0  # For LoRA and efficient fine-tuning
tensorboard>=2.15.0  # For training visualization

# Optional: For quantization support (8-bit/4-bit models)
bitsandbytes>=0.41.0

# Optional: For specific tokenizers
# sentencepiece>=0.1.99
# protobuf>=3.20.0